Create animate module
