+++ lib/anime.esm.js
@@ -33,7 +33,8 @@ const defaultTweenSettings = {
 // Global settings
 
 const settings = {
-  suspendWhenDocumentHidden: true
+  speed: 1,
+  suspendWhenDocumentHidden: true,
 };
 
 // Transforms
@@ -681,33 +682,13 @@ function getPath(path, percent) {
   }
 }
 
-function getPathPoint(pathEl, progress, offset = 0) {
-  const length = progress + offset >= 1 ? progress + offset : 0;
-  return pathEl.getPointAtLength(length);
-}
-
-function getPathProgress(pathObject, progress, isPathTargetInsideSVG) {
-  const pathEl = pathObject.el;
-  const parentSvg = getParentSvg(pathEl, pathObject.svg);
-  const p = getPathPoint(pathEl, progress, 0);
-  const p0 = getPathPoint(pathEl, progress, -1);
-  const p1 = getPathPoint(pathEl, progress, +1);
-  const scaleX = isPathTargetInsideSVG ? 1 : parentSvg.w / parentSvg.vW;
-  const scaleY = isPathTargetInsideSVG ? 1 : parentSvg.h / parentSvg.vH;
-  switch (pathObject.property) {
-    case 'x': return (p.x - parentSvg.x) * scaleX;
-    case 'y': return (p.y - parentSvg.y) * scaleY;
-    case 'angle': return Math.atan2(p1.y - p0.y, p1.x - p0.x) * 180 / pi;
-  }
-}
-
-function parseTargets(targets) {
+function parseTargets$1(targets) {
   const targetsArray = targets ? (flattenArray(is.arr(targets) ? targets.map(toArray) : toArray(targets))) : [];
   return filterArray(targetsArray, (item, pos, self) => self.indexOf(item) === pos);
 }
 
 function getAnimatables(targets) {
-  const parsed = parseTargets(targets);
+  const parsed = parseTargets$1(targets);
   return parsed.map((t, i) => {
     return {
       target: t,
@@ -982,10 +963,7 @@ if (isBrowser) {
   document.addEventListener('visibilitychange', handleVisibilityChange);
 }
 
-// Public Instance
-
-function anime(params = {}) {
-
+function animate(params = {}) {
   let startTime = 0, lastTime = 0, now = 0;
   let children, childrenLength = 0;
   let resolve = null;
@@ -1014,7 +992,7 @@ function anime(params = {}) {
 
   function resetTime() {
     startTime = 0;
-    lastTime = adjustTime(instance.currentTime) * (1 / anime.speed);
+    lastTime = adjustTime(instance.currentTime) * (1 / settings.speed);
   }
 
   function seekChild(time, child) {
@@ -1190,7 +1168,7 @@ function anime(params = {}) {
   instance.tick = function(t) {
     now = t;
     if (!startTime) startTime = now;
-    setInstanceProgress((now + (lastTime - startTime)) * anime.speed);
+    setInstanceProgress((now + (lastTime - startTime)) * settings.speed);
   };
 
   instance.seek = function(time) {
@@ -1229,10 +1207,11 @@ function anime(params = {}) {
 
   instance.reset();
 
-  if (instance.autoplay) instance.play();
+  if (instance.autoplay) {
+    instance.play();
+  }
 
   return instance;
-
 }
 
 // Remove targets from animation
@@ -1245,7 +1224,7 @@ function removeTargetsFromAnimations(targetsArray, animations) {
   }
 }
 
-function removeTargetsFromInstance(targetsArray, instance) {
+function removeTargetsFromInstance$1(targetsArray, instance) {
   const animations = instance.animations;
   const children = instance.children;
   removeTargetsFromAnimations(targetsArray, animations);
@@ -1259,10 +1238,10 @@ function removeTargetsFromInstance(targetsArray, instance) {
 }
 
 function removeTargetsFromActiveInstances(targets) {
-  const targetsArray = parseTargets(targets);
+  const targetsArray = parseTargets$1(targets);
   for (let i = activeInstances.length; i--;) {
     const instance = activeInstances[i];
-    removeTargetsFromInstance(targetsArray, instance);
+    removeTargetsFromInstance$1(targetsArray, instance);
   }
 }
 
@@ -1317,7 +1296,7 @@ function stagger(val, params = {}) {
 // Timeline
 
 function timeline(params = {}) {
-  let tl = anime(params);
+  let tl = animate(params);
   tl.duration = 0;
   tl.add = function(instanceParams, timelineOffset) {
     const tlIndex = activeInstances.indexOf(tl);
@@ -1332,7 +1311,7 @@ function timeline(params = {}) {
     insParams.timelineOffset = is.und(timelineOffset) ? tlDuration : getRelativeValue(timelineOffset, tlDuration);
     passThrough(tl);
     tl.seek(insParams.timelineOffset);
-    const ins = anime(insParams);
+    const ins = animate(insParams);
     passThrough(ins);
     children.push(ins);
     const timings = getInstanceTimings(children, params);
@@ -1364,6 +1343,8 @@ function setTargetsValue(targets, properties) {
     }
   });
 }
+
+const anime = animate;
 
 anime.version = '3.3.0';
 anime.speed = 1;
